+++ src/HappyZu.CloudStore.Core/Trip/ETicketManager.cs
@@ -24,7 +24,7 @@ namespace HappyZu.CloudStore.Trip
             await _eTicketRepository.UpdateAsync(eTicket);
         }
 
-        public async Task<ETicket> GetETicketBySerialNoAndHashAsync(int serailNo, string hash)
+        public async Task<ETicket> GetETicketBySerialNoAndHashAsync(long serailNo, string hash)
         {
             return await _eTicketRepository.FirstOrDefaultAsync(
                         eTicket => eTicket.SerialNo == serailNo && eTicket.Hash == hash);
